add Fedora OS check to all install scripts

+++ fonts/install.sh
@@ -1,7 +1,14 @@
+#!/bin/bash
+if [ "$OS_NAME" != "Fedora" ]; then
+	exit 0
+fi
+
 echo "[INFO] Installing Fonts"
-# Fonts from https://github.com/google/fonts
+
 if [ -e ~/.fonts ]; then
 	echo "[INFO] removing old ~/.fonts directory"
 	rm ~/.fonts
 fi
+
+# Fonts from https://github.com/google/fonts
 ln -s $(pwd)/fonts/.fonts ~/.fonts
